Remove ToExpr.derived and FromExpr.derived in favor of ToExprFactory/FromExprFactory
ToExprFactory/FromExprFactory cover everything ToExpr.derived/FromExpr.derived
did (generic and concrete types) plus generic type parameters and fields
whose ToExpr/FromExpr instance needs Type[_] without requiring Quotes at the
call site, so the two parallel mechanisms are no longer needed. The
inline-based derivedProduct/derivedSum bodies are now self-contained again
(no shared InternalUtils helpers), since sharing them via a generic method
call broke case-object sum dispatch.

Also restores the LowPriorityToExpr/LowPriorityFromExpr split for the
ToExprFactory/FromExprFactory bridge given, otherwise it's ambiguous with
ValueOfToExpr for singleton types.
